# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Workload Evaluation API
  version: 0.1.0
  description: >
    本地工作量评估系统 API。响应统一为 { code, message, data, requestId? }。

    **鉴权（以运行代码为准）**：当前实现使用 JWT，请求头 `Authorization: Bearer <token>`。
    本文档中仍出现的 `X-Role` 参数为历史草稿，**运行时不再作为权限依据**。

    **未在本文档中展开的已实现端点**（完整列表见仓库
    `03_技术设计/系统演进/实现与文档对齐说明.md`）包括但不限于：
    `/api/v1/auth/*`、`/api/v1/versions/*`、`/api/v1/system/*`、
    `/api/v1/ai/parse-basic-info`、`/api/v1/ai/company-profile-summary`、`/api/v1/ai/kimi-assessment/preview`、`/api/v1/ai/kimi-assessment/export-markdown`、
    `/api/v1/teams/*`、`GET /api/v1/wbs`；
    下载路径为 `GET /downloads/{fileName}`（非 `/api/v1` 前缀）。

    **`/api/v1/agent/*` 说明**：当前已挂载 V1 最小基座 `POST /api/v1/agent/chat`。
    该接口返回统一 JSON 事件数组；抽取、PDF、真实归档工具和前端工作台仍在后续批次接线。
servers:
  - url: http://localhost:3000
    description: Local API server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/ai/parse-basic-info:
    post:
      tags: [AI]
      summary: AI 解析需求导入文件（Excel）
      operationId: parseBasicInfo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 解析成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseBasicInfoResponse"
        "400":
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 无权限
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/ai/parse-basic-info/stream:
    post:
      tags: [AI]
      summary: AI 解析需求导入文件（SSE 流式）
      operationId: parseBasicInfoStream
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: SSE 流
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 无权限
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "登录 `POST /api/v1/auth/login` 获取 token，请求头 `Authorization: Bearer <token>`"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        requestId:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              reason:
                type: string
            required: [field, reason]
      required: [code, message, details, requestId]
...
    ParseBasicInfoResponse:
      type: object
      properties:
        code:
          type: integer
          const: 0
        message:
          type: string
          const: ok
        data:
          type: object
          properties:
            basicInfo:
              type: object
            requirementImportData:
              type: object
            sourceSheets:
              type: array
              items:
                type: string
            model:
              type: string
            mode:
              type: string
              enum: [model, rule_fallback]
            fallbackReason:
              type: string
            rawContent:
              type: string
          required: [basicInfo, sourceSheets, model, mode]
      required: [code, message, data]
...